requirements.txt
pyspark==3.5.1
boto3==1.34.89
python-dotenv==1.0.1
pandas==2.2.2
